#1334 clean up dropdown
