+++ lib/elementWrapper/elementWrapperFactory.js
@@ -24,8 +24,8 @@ const createElementWrapper = (elementType, query, attrValuePairs, _options, ...a
       return new RadioButtonWrapper(selector, options, description);
     case 'Checkbox':
       return new CheckBoxWrapper(selector, options, description);
-    // case 'dropDown':
-    //   return new DropDownWrapper(get, description, getIfExists);
+    case 'DropDown':
+      return new DropDownWrapper(selector, options, description);
     // case 'colorField':
     //   return new ColorWrapper(get, description, getIfExists);
     // case 'rangeType':
