+++ lib/taiko.js
@@ -2578,41 +2578,7 @@ module.exports.dropDown = dropDown;
 
 function dropDown(attrValuePairs, _options, ...args) {
   validate();
-  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
-  function getDropDownElementWithLabel(label) {
-    const fileField = [];
-    function checkAndPushElement(elem) {
-      if (elem.tagName && elem.tagName.toLowerCase() == 'select') {
-        fileField.push(elem);
-      }
-    }
-    const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
-      return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
-    });
-    for (let matchingLabel of matchingLabels) {
-      const labelFor = matchingLabel.getAttribute('for');
-      if (labelFor) {
-        //check label with attribute for
-        const labelForElement = document.getElementById(labelFor);
-        checkAndPushElement(labelForElement);
-      } else {
-        // check child node of label tag
-        matchingLabel.childNodes.forEach((elem) => {
-          checkAndPushElement(elem);
-        });
-      }
-    }
-    return fileField;
-  }
-  const get = getElementGetter(
-    selector,
-    async () =>
-      await $function(getDropDownElementWithLabel, selector.label, options.selectHiddenElements),
-    'select',
-    options.selectHiddenElements,
-  );
-  const description = desc(selector, 'label', 'DropDown');
-  return createElementWrapper(get, description, 'dropDown');
+  return createElementWrapper('DropDown', 'label', attrValuePairs, _options, ...args);
 }
 
 /**
